+++ Src/XmlDocInspections.Plugin.Tests/Integrative/BasicTests.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using JetBrains.Application.Settings;
+﻿using JetBrains.Application.Settings;
 using NUnit.Framework;
 using XmlDocInspections.Plugin.Settings;