+++ Src/XmlDocInspections.Plugin.Tests/test/data/QuickFixes/AddDocCommentFix/Class.cs
@@ -1,0 +1,6 @@
+﻿namespace XmlDocInspections.Sample.QuickFixes.AddDocCommentFix
+{
+    public class Class{caret}
+    {
+    }
+}