+++ Src/XmlDocInspections.Plugin.Tests/test/data/QuickFixes/AddDocCommentFix/Field.cs
@@ -1,0 +1,7 @@
+﻿namespace XmlDocInspections.Sample.QuickFixes.AddDocCommentFix
+{
+    public class Field
+    {
+        public string SomeField{caret};
+    }
+}