+++ Src/XmlDocInspections.Plugin.Tests/test/data/QuickFixes/AddDocCommentFix/MethodWithAdditionalElementsForDocTemplate.cs
@@ -1,0 +1,10 @@
+﻿namespace XmlDocInspections.Sample.QuickFixes.AddDocCommentFix
+{
+    public class MethodWithAdditionalElementsForDocTemplate
+    {
+        public string Method{caret}<T>(string a, params object[] p)
+        {
+            return "";
+        }
+    }
+}