+++ Src/XmlDocInspections.Plugin.Tests/test/data/QuickFixes/AddDocCommentFix/Property.cs
@@ -1,0 +1,7 @@
+﻿namespace XmlDocInspections.Sample.QuickFixes.AddDocCommentFix
+{
+    public class Property
+    {
+        public string SomeProperty{caret} { get; set; }
+    }
+}