+++ Src/XmlDocInspections.Plugin.Tests/test/data/QuickFixes/AddDocCommentFix/SimpleMethod.cs
@@ -1,0 +1,9 @@
+﻿namespace XmlDocInspections.Sample.QuickFixes.AddDocCommentFix
+{
+    public class SimpleMethod
+    {
+        public void Met{caret}hod()
+        {
+        }
+    }
+}